+++ source/programs/conversion/grace/graceL1a2SatelliteTracking.cpp
@@ -1,6 +1,6 @@
 /***********************************************/
 /**
-* @file graceL1A2SatelliteTracking.cpp
+* @file graceL1a2SatelliteTracking.cpp
 *
 * @brief Read GRACE L1A data.
 *
@@ -16,7 +16,6 @@ static const char *docstring = R"(
 This program converts Level-1A satellite tracking data (KBR1A) to the GROOPS instrument file format.
 The GRACE Level-1A format is described in \verb|GRACEiolib.h| given at
 \url{http://podaac-tools.jpl.nasa.gov/drive/files/allData/grace/sw/GraceReadSW_L1_2010-03-31.tar.gz}.
-Multiple \config{inputfile}s must be given in the correct time order.
 The output is one arc of satellite data which can include data gaps.
 To split the arc in multiple gap free arcs use \program{InstrumentSynchronize}.
 )";
@@ -31,24 +30,24 @@ To split the arc in multiple gap free arcs use \program{InstrumentSynchronize}.
 
 /** @brief Read Level-1A GRACE data.
 * @ingroup programsConversionGroup */
-class GraceL1A2SatelliteTracking
+class GraceL1a2SatelliteTracking
 {
 public:
   void run(Config &config, Parallel::CommunicatorPtr comm);
 };
 
-GROOPS_REGISTER_PROGRAM(GraceL1A2SatelliteTracking, SINGLEPROCESS, "read GRACE L1A data (KBR1A)", Conversion, Grace, Instrument)
+GROOPS_REGISTER_PROGRAM(GraceL1a2SatelliteTracking, SINGLEPROCESS, "read GRACE L1A data (KBR1A)", Conversion, Grace, Instrument)
 
 /***********************************************/
 
-void GraceL1A2SatelliteTracking::run(Config &config, Parallel::CommunicatorPtr /*comm*/)
+void GraceL1a2SatelliteTracking::run(Config &config, Parallel::CommunicatorPtr /*comm*/)
 {
   try
   {
     FileName fileNameSst;
     std::vector<FileName> fileNameIn;
 
-    readConfig(config, "outputfileSatelliteTracking", fileNameSst, Config::OPTIONAL, "", "MISCVALUES(seconds, microSeconds, ant_id, K_phase, Ka_phase, K_SNR, Ka_SNR)");
+    readConfig(config, "outputfileSatelliteTracking", fileNameSst, Config::OPTIONAL, "", "MISCVALUES(ant_id, K_phase, Ka_phase, K_SNR, Ka_SNR)");
     readConfig(config, "inputfile",                   fileNameIn,  Config::MUSTSET,  "", "KBR1A");
     if(isCreateSchema(config)) return;
 
@@ -103,23 +102,15 @@ void GraceL1A2SatelliteTracking::run(Config &config, Parallel::CommunicatorPtr /
           break;
         }
 
-        while(microSeconds >= 1'000'000)
-        {
-          seconds      += 1;
-          microSeconds -= 1'000'000;
-        }
-
         if((ant_id == 9) || (ant_id == 11) || (ant_id == -11))
         {
           MiscValuesEpoch epoch(7);
           epoch.time      = mjd2time(51544.5) + seconds2time(seconds) + seconds2time(microSeconds*1e-6);
-          epoch.values(0) = seconds;
-          epoch.values(1) = microSeconds;
-          epoch.values(2) = ant_id;
-          epoch.values(3) = K_phase;
-          epoch.values(4) = Ka_phase;
-          epoch.values(5) = K_SNR;
-          epoch.values(6) = Ka_SNR;
+          epoch.values(0) = ant_id;
+          epoch.values(1) = K_phase;
+          epoch.values(2) = Ka_phase;
+          epoch.values(3) = K_SNR;
+          epoch.values(4) = Ka_SNR;
           arc.push_back(epoch);
         }
       } // for(idEpoch)
@@ -127,7 +118,16 @@ void GraceL1A2SatelliteTracking::run(Config &config, Parallel::CommunicatorPtr /
 
     // =============================================
 
+
+    logStatus<<"sort epochs"<<Log::endl;
     arc.sort();
+
+    logStatus<<"eliminate duplicates"<<Log::endl;
+    const UInt oldSize = arc.size();
+    arc.removeDuplicateEpochs(TRUE/*keepFirst*/);
+    if(arc.size() < oldSize)
+      logInfo<<" "<<oldSize-arc.size()<<" duplicates removed!"<<Log::endl;
+
     Arc::printStatistics(arc);
     if(arc.size() == 0)
       return;