+++ source/programs/conversion/grace/graceL1b2ClockOffset.cpp
@@ -13,7 +13,7 @@
 // Latex documentation
 #define DOCSTRING docstring
 static const char *docstring = R"(
-This program converts clock data (CLK1B or LLK1B) from the GRACE SDS format into \file{instrument file (CLOCK)}{instrument}.
+This program converts clock data (CLK1B or LLK1B) from the GRACE SDS format into \file{instrument file (MISCVALUE)}{instrument}.
 For further information see \program{GraceL1b2Accelerometer}.
 )";
 
@@ -27,24 +27,24 @@ For further information see \program{GraceL1b2Accelerometer}.
 
 /** @brief Read GRACE L1B data.
 * @ingroup programsConversionGroup */
-class GraceL1b2Clock
+class GraceL1b2ClockOffset
 {
 public:
   void run(Config &config, Parallel::CommunicatorPtr comm);
 };
 
-GROOPS_REGISTER_PROGRAM(GraceL1b2Clock, SINGLEPROCESS, "read GRACE L1B data (CLK1B or LLK1B)", Conversion, Grace, Instrument)
+GROOPS_REGISTER_PROGRAM(GraceL1b2ClockOffset, SINGLEPROCESS, "read GRACE L1B data (CLK1B or LLK1B)", Conversion, Grace, Instrument)
 
 /***********************************************/
 
-void GraceL1b2Clock::run(Config &config, Parallel::CommunicatorPtr /*comm*/)
+void GraceL1b2ClockOffset::run(Config &config, Parallel::CommunicatorPtr /*comm*/)
 {
   try
   {
     FileName              fileNameOut;
     std::vector<FileName> fileNameIn;
 
-    readConfig(config, "outputfileClock", fileNameOut, Config::MUSTSET, "", "CLOCK");
+    readConfig(config, "outputfileClock", fileNameOut, Config::MUSTSET, "", "MISCVALUE");
     readConfig(config, "inputfile",       fileNameIn,  Config::MUSTSET, "", "CLK1B or LLK1B");
     if(isCreateSchema(config)) return;
 
@@ -58,7 +58,6 @@ void GraceL1b2Clock::run(Config &config, Parallel::CommunicatorPtr /*comm*/)
       UInt numberOfRecords;
       FileInGrace file(fileNameIn.at(idFile), numberOfRecords);
 
-      Bool dummy = FALSE;
       for(UInt idEpoch=0; idEpoch<numberOfRecords; idEpoch++)
       {
         Int32             seconds;
@@ -82,55 +81,9 @@ void GraceL1b2Clock::run(Config &config, Parallel::CommunicatorPtr /*comm*/)
         if(arc.size() && (time <= arc.back().time))
           logWarning<<"epoch("<<time.dateTimeStr()<<") <= last epoch("<<arc.back().time.dateTimeStr()<<")"<<Log::endl;
 
-        // invalid time periods
-        UInt quality = 0;
-        if(qualflg & (1 << 0)) quality = 1;
-        if(qualflg & (1 << 1)) quality = 2;
-
-        // boundaries - days
-        if((idEpoch==0) && (quality == 2))
-          quality = 0;
-        if((arc.size()) && (idEpoch==1) && (time == arc.back().time))
-          continue;
-
-        if((arc.size()) && (idEpoch==numberOfRecords-1) && (quality == 1) && (time == arc.back().time))
-          continue;
-
-        ClockEpoch epoch;
-        epoch.time        = time;
-        epoch.rcvTime     = seconds;
-        epoch.epsTime     = eps_time;
-        epoch.epsError    = eps_err;
-        epoch.epsDrift    = eps_drift;
-        epoch.driftError  = drift_err;
-        epoch.qualityFlag = quality;
-        //arc.push_back(epoch);
-        //epoch.values = {Double(clock_id), eps_time, eps_err, eps_drift, drift_err, Double(qualflg)};
-
-         // delete invalid periods
-        if(quality == 1)
-        {
-          dummy = TRUE;
-          arc.push_back(epoch);
-          continue;
-        }
-        if((dummy) && (quality != 2))
-          continue;
-        if((dummy) && (quality == 2))
-        {
-          dummy = FALSE;
-          if((arc.size()) && (epoch.time == arc.back().time) && (arc.back().data().at(1) != 0))
-            continue;
-          if((arc.size()) && (epoch.time == arc.back().time) && (arc.back().data().at(1) == 0))
-          {
-            arc.remove(arc.size()-1,1);
-            if(epoch.epsTime != 0)
-            {
-              arc.push_back(epoch);
-              continue;
-            }
-          }
-        }
+        MiscValueEpoch epoch;
+        epoch.time  = time;
+        epoch.value = eps_time;
         arc.push_back(epoch);
       } // for(idEpoch)
     } // for(idFile)