+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Bugfix:           Plot: various fixes for GMT Versions >6.1.
 - Bugfix:           GriddedData: unwrap longitude before computing area elements.
 - Bugfix:           InstrumentResample/polynomial: corrected interpolation vs. extrapolation.
+- Other:            Reworked of GRACE-(FO) L1A conversion.
 
 # Release 2022-07-28
 - New programs:     GriddedData2GriddedDataTimeSeries and GriddedDataTimeSeries2GriddedData.